Change rtol, test with profiling
